+++ src/main/scala/com/redis/cluster/RedisCluster.scala
@@ -24,11 +24,11 @@ import serialization._
  * redis-rb implements a regex based trick to achieve key-tagging. Here is the technique
  * explained in redis FAQ:
  * <i>
- * A key tag is a special pattern inside a key that, if preset, is the only part of the key 
- * hashed in order to select the server for this key. For example in order to hash the key 
- * "foo" I simply perform the CRC32 checksum of the whole string, but if this key has a 
- * pattern in the form of the characters {...} I only hash this substring. So for example 
- * for the key "foo{bared}" the key hashing code will simply perform the CRC32 of "bared". 
+ * A key tag is a special pattern inside a key that, if preset, is the only part of the key
+ * hashed in order to select the server for this key. For example in order to hash the key
+ * "foo" I simply perform the CRC32 checksum of the whole string, but if this key has a
+ * pattern in the form of the characters {...} I only hash this substring. So for example
+ * for the key "foo{bared}" the key hashing code will simply perform the CRC32 of "bared".
  * This way using key tags you can ensure that related keys will be stored on the same Redis
  * instance just using the same key tag for all this keys. Redis-rb already implements key tags.
  * </i>
@@ -69,7 +69,7 @@ abstract class RedisCluster(hosts: String*) extends RedisClient {
   val POINTS_PER_SERVER = 160 // default in libmemcached
 
   // instantiating a cluster will automatically connect participating nodes to the server
-  val clients = hosts.toList.map {h => 
+  val clients = hosts.toList.map {h =>
     val hp = h.split(":")
     new RedisClientPool(hp(0), hp(1).toInt)
   }
@@ -95,7 +95,7 @@ abstract class RedisCluster(hosts: String*) extends RedisClient {
   override def keys[A](pattern: Any = "*")(implicit format: Format, parse: Parse[A]) =
     Some(hr.cluster.toList.map(_.withClient(_.keys[A](pattern))).flatten.flatten)
 
-  def onAllConns[T](body: RedisClient => T) = 
+  def onAllConns[T](body: RedisClient => T) =
     hr.cluster.map(p => p.withClient { client => body(client) }) // .forall(_ == true)
 
   override def flushdb = onAllConns(_.flushdb) forall(_ == true)
@@ -138,9 +138,9 @@ abstract class RedisCluster(hosts: String*) extends RedisClient {
   override def getset[A](key: Any, value: Any)(implicit format: Format, parse: Parse[A]) = nodeForKey(key).getset(key, value)
   override def setnx(key: Any, value: Any)(implicit format: Format) = nodeForKey(key).setnx(key, value)
   override def incr(key: Any)(implicit format: Format) = nodeForKey(key).incr(key)
-  override def incrby(key: Any, increment: Int)(implicit format: Format) = nodeForKey(key).incrby(key, increment)
+  override def incrby(key: Any, increment: Long)(implicit format: Format) = nodeForKey(key).incrby(key, increment)
   override def decr(key: Any)(implicit format: Format) = nodeForKey(key).decr(key)
-  override def decrby(key: Any, increment: Int)(implicit format: Format) = nodeForKey(key).decrby(key, increment)
+  override def decrby(key: Any, increment: Long)(implicit format: Format) = nodeForKey(key).decrby(key, increment)
 
   override def mget[A](key: Any, keys: Any*)(implicit format: Format, parse: Parse[A]): Option[List[Option[A]]] = {
     val keylist = (key :: keys.toList)
@@ -168,14 +168,14 @@ abstract class RedisCluster(hosts: String*) extends RedisClient {
   override def lrem(key: Any, count: Int, value: Any)(implicit format: Format) = nodeForKey(key).lrem(key, count, value)
   override def lpop[A](key: Any)(implicit format: Format, parse: Parse[A]) = nodeForKey(key).lpop[A](key)
   override def rpop[A](key: Any)(implicit format: Format, parse: Parse[A]) = nodeForKey(key).rpop[A](key)
-  override def rpoplpush[A](srcKey: Any, dstKey: Any)(implicit format: Format, parse: Parse[A]) = 
+  override def rpoplpush[A](srcKey: Any, dstKey: Any)(implicit format: Format, parse: Parse[A]) =
     inSameNode(srcKey, dstKey) {n => n.rpoplpush[A](srcKey, dstKey)}
 
   private def inSameNode[T](keys: Any*)(body: RedisClient => T)(implicit format: Format): T = {
     val nodes = keys.toList.map(nodeForKey(_))
     nodes.forall(_ == nodes.head) match {
       case true => body(nodes.head)  // all nodes equal
-      case _ => 
+      case _ =>
         throw new UnsupportedOperationException("can only occur if both keys map to same node")
     }
   }
@@ -187,28 +187,28 @@ abstract class RedisCluster(hosts: String*) extends RedisClient {
   override def srem(key: Any, value: Any, values: Any*)(implicit format: Format): Option[Int] = nodeForKey(key).srem(key, value, values:_*)
   override def spop[A](key: Any)(implicit format: Format, parse: Parse[A]) = nodeForKey(key).spop[A](key)
 
-  override def smove(sourceKey: Any, destKey: Any, value: Any)(implicit format: Format) = 
+  override def smove(sourceKey: Any, destKey: Any, value: Any)(implicit format: Format) =
     inSameNode(sourceKey, destKey) {n => n.smove(sourceKey, destKey, value)}
 
   override def scard(key: Any)(implicit format: Format) = nodeForKey(key).scard(key)
   override def sismember(key: Any, value: Any)(implicit format: Format) = nodeForKey(key).sismember(key, value)
 
-  override def sinter[A](key: Any, keys: Any*)(implicit format: Format, parse: Parse[A]) = 
+  override def sinter[A](key: Any, keys: Any*)(implicit format: Format, parse: Parse[A]) =
     inSameNode((key :: keys.toList): _*) {n => n.sinter[A](key, keys: _*)}
 
-  override def sinterstore(key: Any, keys: Any*)(implicit format: Format) = 
+  override def sinterstore(key: Any, keys: Any*)(implicit format: Format) =
     inSameNode((key :: keys.toList): _*) {n => n.sinterstore(key, keys: _*)}
 
-  override def sunion[A](key: Any, keys: Any*)(implicit format: Format, parse: Parse[A]) = 
+  override def sunion[A](key: Any, keys: Any*)(implicit format: Format, parse: Parse[A]) =
     inSameNode((key :: keys.toList): _*) {n => n.sunion[A](key, keys: _*)}
 
-  override def sunionstore(key: Any, keys: Any*)(implicit format: Format) = 
+  override def sunionstore(key: Any, keys: Any*)(implicit format: Format) =
     inSameNode((key :: keys.toList): _*) {n => n.sunionstore(key, keys: _*)}
 
-  override def sdiff[A](key: Any, keys: Any*)(implicit format: Format, parse: Parse[A]) = 
+  override def sdiff[A](key: Any, keys: Any*)(implicit format: Format, parse: Parse[A]) =
     inSameNode((key :: keys.toList): _*) {n => n.sdiff[A](key, keys: _*)}
 
-  override def sdiffstore(key: Any, keys: Any*)(implicit format: Format) = 
+  override def sdiffstore(key: Any, keys: Any*)(implicit format: Format) =
     inSameNode((key :: keys.toList): _*) {n => n.sdiffstore(key, keys: _*)}
 
   override def smembers[A](key: Any)(implicit format: Format, parse: Parse[A]) = nodeForKey(key).smembers(key)
@@ -221,14 +221,14 @@ abstract class RedisCluster(hosts: String*) extends RedisClient {
   /**
    * SortedSetOperations
    */
-  override def zadd(key: Any, score: Double, member: Any, scoreVals: (Double, Any)*)(implicit format: Format) = 
+  override def zadd(key: Any, score: Double, member: Any, scoreVals: (Double, Any)*)(implicit format: Format) =
     nodeForKey(key).zadd(key, score, member, scoreVals:_*)
-  override def zrem(key: Any, member: Any, members: Any*)(implicit format: Format): Option[Int] = 
+  override def zrem(key: Any, member: Any, members: Any*)(implicit format: Format): Option[Int] =
     nodeForKey(key).zrem(key, member, members)
   override def zincrby(key: Any, incr: Double, member: Any)(implicit format: Format) = nodeForKey(key).zincrby(key, incr, member)
   override def zcard(key: Any)(implicit format: Format) = nodeForKey(key).zcard(key)
   override def zscore(key: Any, element: Any)(implicit format: Format) = nodeForKey(key).zscore(key, element)
-  override def zrange[A](key: Any, start: Int = 0, end: Int = -1, sortAs: SortOrder )(implicit format: Format, parse: Parse[A]) = 
+  override def zrange[A](key: Any, start: Int = 0, end: Int = -1, sortAs: SortOrder )(implicit format: Format, parse: Parse[A]) =
     nodeForKey(key).zrange[A](key, start, end, sortAs)
   override def zrangeWithScore[A](key: Any, start: Int = 0, end: Int = -1, sortAs: SortOrder = ASC)(implicit format: Format, parse: Parse[A]) =
     nodeForKey(key).zrangeWithScore[A](key, start, end, sortAs)
@@ -245,7 +245,7 @@ abstract class RedisCluster(hosts: String*) extends RedisClient {
   override def hget[A](key: Any, field: Any)(implicit format: Format, parse: Parse[A]) = nodeForKey(key).hget[A](key, field)
   override def hmset(key: Any, map: Iterable[Product2[Any, Any]])(implicit format: Format) = nodeForKey(key).hmset(key, map)
   override def hmget[K,V](key: Any, fields: K*)(implicit format: Format, parseV: Parse[V]) = nodeForKey(key).hmget[K,V](key, fields:_*)
-  override def hincrby(key: Any, field: Any, value: Int)(implicit format: Format) = nodeForKey(key).hincrby(key, field, value)
+  override def hincrby(key: Any, field: Any, value: Long)(implicit format: Format) = nodeForKey(key).hincrby(key, field, value)
   override def hexists(key: Any, field: Any)(implicit format: Format) = nodeForKey(key).hexists(key, field)
   override def hdel(key: Any, field: Any, fields: Any*)(implicit format: Format): Option[Int] = nodeForKey(key).hdel(key, field, fields:_*)
   override def hlen(key: Any)(implicit format: Format) = nodeForKey(key).hlen(key)
